% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_participation_rate.R
\name{indicator_participation_rate}
\alias{indicator_participation_rate}
\title{Indicator: Participation rate}
\usage{
indicator_participation_rate(dt, stratum.col.nms = NULL, subset = NULL)
}
\arguments{
\item{dt}{screening data in the CDM format defined in the EUCanScreen project
(\code{data.table})}

\item{stratum.col.nms}{names of columns used for stratification as a character vector
(default \code{screening_year})}

\item{subset}{a logical expression indicating the rows to keep}
}
\value{
A \code{data.table} of ...
}
\description{
The proportion of invited individuals who have undergone a screening
test within a given time frame following an active invitation.
}
\details{
\subsection{Calculation}{
}

\subsection{Dimensions/ Disaggregation}{

\describe{
}
}

\subsection{Further Specifications}{

\describe{
}
}

\subsection{Relevant cancer sites}{

\describe{
}
}

\subsection{Reference standard}{

N/A
}
}
\examples{
# example code

}
