% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_test_result.R
\name{indicator_test_result}
\alias{indicator_test_result}
\title{Indicator: Test result}
\usage{
indicator_test_result(
  dt,
  result.col.nm = "primary_test_result_factor",
  stratum.col.nms = NULL,
  subset = NULL,
  na.rm = TRUE
)
}
\arguments{
\item{dt}{screening data in the CDM format defined in the EUCanScreen project
(\code{data.table})}

\item{result.col.nm}{name of the categorical result column as a character (default
\code{"primary_result_factor"})}

\item{stratum.col.nms}{names of columns used for stratification as a character vector
(default \code{NULL})}

\item{subset}{a logical expression indicating the rows to keep}

\item{na.rm}{a logical expression; if \code{TRUE}, drops rows that have \code{NA} values in
column result.col.nm (default \code{FALSE})}
}
\value{
A \code{data.table} of counts and percentages of each result category,
further stratified by the columns defined in \code{stratum.col.nms}.
}
\description{
The results of the screening test.
Monitoring the different results of the screening test has many benefits.
The number of positive screening tests determines the cancer diagnostic
assessment workload and cancer detection rate and is important for program
planning and quality assurance. Any substantial changes to these results
over time will require investigation.
}
\details{
\subsection{Calculation}{

N = Number of tests with either positive, negative, inadequate, or
borderline results \cr
D = Total number of screening tests within the program
}

\subsection{Dimensions/ Disaggregation}{

\describe{
\item{\emph{Minimum_Requirements:}}{Sex, Age Group, Screening Test Type}
\item{\emph{Desirable:}}{Education, Income, Migration Status, Country of Birth,
Country of Citizenship, Country of Parents Birth, Employment Status,
Disability Status, Screening Invitation Round, Previous Screening Attendance}
\item{\emph{Future:}}{Risk Status, LGBTIQA+ status, Geographic Location, Deprivation Area,
Vaccination status}
}
}

\subsection{Further Specifications}{

\describe{
\item{\emph{Breast:}}{Positive, negative, inadequate, unknown --
Management: Positive: referral to further assessment. Negative: not
referred to further assessment. Inadequate: test result inadequate, retest
needed. Unknown: test result not know.}
\item{\emph{Cervical:}}{Positive, borderline, negative, inadequate, unknown --
Management: Positive: referral to triage cytology or colposcopy.
Borderline: Recommendation to follow up testing Negative: not referred to
colposcopy. Inadequate: test result inadequate, retest needed. Unknown:
test result not known.}
\item{\emph{Colorectal:}}{Positive, negative, inadequate, unknown --
Management: Positive: referral to colonoscopy (diagnostic confirmation).
Negative: not referred to colonoscopy (test result known). Inadequate:
test result inadequate, retest needed. Unknown: test result not known.
Note: result depending on the test system.}
\item{\emph{Lung:}}{Positive, negative, indeterminate, inadequate, unknown --
Management: Positive: referral to further assessment. Negative: regular
follow-up LDCT (usually after 12 months). Indeterminate: need for early
rescreen (LDCT already after 3 or 6 months). Inadequate: test result
inadequate, retest needed. Unknown: test result not known.}
}
}

\subsection{Relevant cancer sites}{

\describe{
\item{\emph{Current:}}{Breast, Cervical, Colorectal, Prostate, Lung}
\item{\emph{Future:}}{}
}
}

\subsection{Reference standard}{

N/A
}
}
\examples{
# example code

}
